projects/python-crawler/crawler.py
"""
彩票数据爬虫模块 - 彩票数据分析系统
"""
import requests
import time
import re
from datetime import datetime, timedelta
from typing import List, Dict, Any, Optional
from bs4 import BeautifulSoup
from loguru import logger
from config import CRAWLER_CONFIG, DATA_SOURCES
from database import DatabaseManager


class LotteryCrawler:
    """彩票数据爬虫"""
    
    def __init__(self):
        """初始化爬虫"""
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': CRAWLER_CONFIG['user_agent']
        })
        self.db = DatabaseManager()
    
    def crawl_dlt_data(self, pages: int = 10) -> List[Dict[str, Any]]:
        """
        爬取大乐透数据
        大乐透格式：前区5个号码(01-35)，后区2个号码(01-12)
        """
        logger.info("开始爬取大乐透数据")
        results = []
        
        try:
            # 这里使用模拟数据，实际项目中需要根据真实网站调整
            for page in range(1, pages + 1):
                # 模拟数据生成
                for i in range(10):  # 每页10条记录
                    draw_number = f"{(page-1)*10 + i + 1:04d}"
                    draw_date = (datetime.now() - timedelta(days=i)).strftime('%Y-%m-%d')
                    
                    # 生成模拟开奖号码
                    front_numbers = sorted([f"{n:02d}" for n in np.random.choice(range(1, 36), 5, replace=False)])
                    back_numbers = sorted([f"{n:02d}" for n in np.random.choice(range(1, 13), 2, replace=False)])
                    
                    result = {
                        'draw_number': draw_number,
                        'draw_date': draw_date,
                        'numbers': {
                            'front': front_numbers,
                            'back': back_numbers
                        },
                        'sales_amount': round(np.random.uniform(1000000, 5000000), 2),
                        'prize_pool': round(np.random.uniform(10000000, 100000000), 2)
                    }
                    results.append(result)
                
                time.sleep(CRAWLER_CONFIG['request_delay'])
                logger.info(f"大乐透第{page}页数据爬取完成")
        
        except Exception as e:
            logger.error(f"大乐透数据爬取失败: {e}")
        
        return results
    
    def crawl_fc3d_data(self, pages: int = 10) -> List[Dict[str, Any]]:
        """
        爬取福彩3D数据
        福彩3D格式：3个号码(000-999)
        """
        logger.info("开始爬取福彩3D数据")
        results = []
        
        try:
            for page in range(1, pages + 1):
                for i in range(10):
                    draw_number = f"{(page-1)*10 + i + 1:04d}"
                    draw_date = (datetime.now() - timedelta(days=i)).strftime('%Y-%m-%d')
                    
                    # 生成模拟开奖号码
                    number = f"{np.random.randint(0, 1000):03d}"
                    
                    result = {
                        'draw_number': draw_number,
                        'draw_date': draw_date,
                        'numbers': {
                            'main': number,
                            'hundred': int(number[0]),
                            'ten': int(number[1]),
                            'unit': int(number[2])
                        },
                        'sales_amount': round(np.random.uniform(500000, 2000000), 2),
                        'prize_pool': round(np.random.uniform(5000000, 50000000), 2)
                    }
                    results.append(result)
                
                time.sleep(CRAWLER_CONFIG['request_delay'])
                logger.info(f"福彩3D第{page}页数据爬取完成")
        
        except Exception as e:
            logger.error(f"福彩3D数据爬取失败: {e}")
        
        return results
    
    def crawl_ssq_data(self, pages: int = 10) -> List[Dict[str, Any]]:
        """
        爬取双色球数据
        双色球格式：红球6个(01-33)，蓝球1个(01-16)
        """
        logger.info("开始爬取双色球数据")
        results = []
        
        try:
            for page in range(1, pages + 1):
                for i in range(10):
                    draw_number = f"{(page-1)*10 + i + 1:04d}"
                    draw_date = (datetime.now() - timedelta(days=i)).strftime('%Y-%m-%d')
                    
                    # 生成模拟开奖号码
                    red_numbers = sorted([f"{n:02d}" for n in np.random.choice(range(1, 34), 6, replace=False)])
                    blue_number = f"{np.random.randint(1, 17):02d}"
                    
                    result = {
                        'draw_number': draw_number,
                        'draw_date': draw_date,
                        'numbers': {
                            'red': red_numbers,
                            'blue': blue_number
                        },
                        'sales_amount': round(np.random.uniform(800000, 4000000), 2),
                        'prize_pool': round(np.random.uniform(15000000, 150000000), 2)
                    }
                    results.append(result)
                
                time.sleep(CRAWLER_CONFIG['request_delay'])
                logger.info(f"双色球第{page}页数据爬取完成")
        
        except Exception as e:
            logger.error(f"双色球数据爬取失败: {e}")
        
        return results
    
    def save_to_database(self, lottery_type_id: int, results: List[Dict[str, Any]]) -> int:
        """保存数据到数据库"""
        success_count = 0
        
        for result in results:
            try:
                success = self.db.insert_lottery_result(
                    lottery_type_id=lottery_type_id,
                    draw_number=result['draw_number'],
                    draw_date=result['draw_date'],
                    numbers=result['numbers'],
                    sales_amount=result.get('sales_amount'),
                    prize_pool=result.get('prize_pool')
                )
                if success:
                    success_count += 1
            except Exception as e:
                logger.error(f"保存数据失败: {e}")
        
        logger.info(f"数据保存完成，成功保存{success_count}条记录")
        return success_count
    
    def crawl_all_data(self) -> Dict[str, int]:
        """爬取所有彩票类型的数据"""
        logger.info("开始爬取所有彩票数据")
        
        # 获取彩票类型
        lottery_types = self.db.get_lottery_types()
        results = {}
        
        for lottery_type in lottery_types:
            try:
                if lottery_type['type_code'] == 'DLT':
                    data = self.crawl_dlt_data()
                    success_count = self.save_to_database(lottery_type['id'], data)
                    results['DLT'] = success_count
                
                elif lottery_type['type_code'] == 'FC3D':
                    data = self.crawl_fc3d_data()
                    success_count = self.save_to_database(lottery_type['id'], data)
                    results['FC3D'] = success_count
                
                elif lottery_type['type_code'] == 'SSQ':
                    data = self.crawl_ssq_data()
                    success_count = self.save_to_database(lottery_type['id'], data)
                    results['SSQ'] = success_count
                
                logger.info(f"{lottery_type['type_name']}数据爬取完成，保存{results.get(lottery_type['type_code'], 0)}条")
                
            except Exception as e:
                logger.error(f"{lottery_type['type_name']}数据爬取失败: {e}")
                results[lottery_type['type_code']] = 0
        
        return results
    
    def close(self):
        """关闭爬虫"""
        self.db.close()


# 添加缺失的导入
import numpy as np 