projects/python-crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2 